Major refactoring: resize and shrink methods
- Resizing function had a bug and was not copying items from the
original array to the new array correctly. Bugs:
(1) we must start with the NEXT item of the nextFront index. If the
current nextFront index is at the end of the array, then it must
wrap to the beginning of the array to 0. The previous code didn't
handle this and simply use +1, which does not have an item.
(2) in the second arraycopy call, we must also calculate headSize
like we calculated tailSize. headSize is the number of elements from
index 0 to the element BEFORE nextBack. We must recalculate the
index of nextBack. If our nextBack is currently at 0, then it means
the last element in our array is at index 7. In this case, we need
to copy the whole array. Since our array is 0-based indexed, to
calculate the size, we must add 1.
- shrink array had a bug: we should only shrink an array when its
use ratio is below 25%. For an array of length 16, we will not
shrink when we have 4 elements. We will only shrink when we have
3. Hence we must change the condition to this.size < this.array.
length * 0.25 (not <=).
- update various tests: get(0) should return turn the first
element of the array