Important bug fix: add multiple files to staging area, and compare current version of the file with the file in the last commit, and reset staging
- Fix mang 3 bugs in this commit
- 1. We should be able to add many files and make them all into
one commit. To do so, we want to keep adding/updating the hashmap
trackedFiles inside the staging object. The problem of past
implementation was that I did not load the staging file before
writing into it, hence I am overwriting the trackedFile over and
over again. Bug fix: instatiate a new staging object for each repo,
save it as soon as we initialize a repo. When adding a file, load
the staging file, update the hashmap, and save it.
- 2. Compare current version of the file should compare the file
in CWD with the same file in the last commit. To do so we should
compare two files at the level of bytes: we want to (1) find out
the blob of the file saved in the hashmap of the current commit,
load its fileContent byte array, then (2) use readContent to read
the file into a byte array, and (3) compare the two byte arrays.
Past implementation only compare the blob SHA1.
- 3. Clear a staging area should not delete the file. Instead,
we "clear" it by reseting the hashmap inside the staging object
to a new Hashmap.